requirement.txt
streamlit
faiss-cpu
numpy
sentence-transformers
google-generativeai
python-dotenv
pip freeze > requirements.txt